Added view for adding wallet
